requirements.txt
transformers==4.18.0
torch==2.1.2
torchvision==0.16.2
torchvision==0.16.2
hydra-core==1.3.2
matplotlib==3.8.2
onnx==1.15.0
onnxruntime==1.16.3

